look forward and backward in the beamline to get the correct offset
if we have a small element at s in the beamline, and we say we want to
to attach a BLM which is longer than that element, if either the
following or proceeding (or both) elements in the beamline have larger
extents then we will get an overlap.  we now look forwards and
backwards to get the greatest tranverse extents, and use those.


Former-commit-id: a0ad35b395267e4178a659bde1699630654d7801