Sample flat in sigma space for a beam halo + test
Why:

If you sample the emittance uniformly (as with the other halo generator), you get a
non-flat PDF in the nsigmas.  This is not ideal for reweighting because often you
want to consider the halo in terms of nsigma, as opposed to the emittance.

Implementation:

Pick a random single particle emittance (action).  This, with alpha and beta Twiss
parameters defines an ellipse.  You can't simply pick a random angle to get a
random point on the ellipse because the ends of the major axes will be
oversampled.  So instead you must sample the arc length uniformly.  Do this by
calculating the angle and arc length pairs for the defined
ellipse (in EllipsePointGenerator ctor and function CumulativeDistances).  The last
one (angle=2pi) is simply the ellipse circumference.  Generate a number uniformly
from 0 to the circumference.  Then interpolate using the previous vector of (angle,
arc_length) pairs to turn this arc length into an angle.  Then you have an action
and an angle, so you have an (x, xp) pair too.  That's it.


Former-commit-id: 615a8e3ade6297723f38c0306adb88776d6fc6e2