Implementation of the option to use a time varying modulation of given field maps in up to three spatial dimensions.
The BDSFieldModulator class has been introduced to take care of the option to modulate a given field map,
which can be electric, magnetic or electro-magnetic and 1D, 2D or 3D (only spatial dimensions).
The modulation can be of sinusodial or cosinusodial form, takes a frequency and a phase or time-offset. The
last two arguments can be calculated out of the other.


Former-commit-id: 91e78eff58bcb0b7c619e496503f7d59e4af7a14